feat: Add minimal sqlite module
start testing table

Issue #21
