+++ tests/database/table.py
@@ -1,0 +1,61 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+#    Copyright (C) 2017, Kai Raphahn <kai.raphahn@laburec.de>
+#
+
+import os
+import unittest
+import unittest.mock as mock
+
+from unittest.mock import Mock
+from bbutil.database import Table
+from bbutil.utils import full_path
+
+from tests.database.helper import set_log, get_sqlite
+
+__all__ = [
+    "TestTable"
+]
+
+
+class TestTable(unittest.TestCase):
+    """Testing class for locking module."""
+
+    def tearDown(self):
+        return
+
+    def test_connect_01(self):
+        _log = set_log()
+        _testfile = full_path("{0:s}/test.sqlite".format(os.getcwd()))
+        _name = "Test"
+
+        if os.path.exists(_testfile) is True:
+            os.remove(_testfile)
+
+        _sqlite = SQLite(filename=_testfile, name="Test", log=_log)
+
+        _check1 = _sqlite.connect()
+        _check2 = os.path.exists(_testfile)
+
+        self.assertEqual(_sqlite.name, _name)
+        self.assertEqual(_sqlite.filename, _testfile)
+        self.assertTrue(_check1)
+        self.assertTrue(_check2)
+        self.assertTrue(_sqlite.is_connected)
+
+        if os.path.exists(_testfile) is True:
+            os.remove(_testfile)
+        return
